# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ALSS-LEPC Adaptive Learning API",
    "description": "Adaptive Learning System with SAINT KT + D3QN RL",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/pipeline/process": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Process Document",
        "description": "Run the full content processing pipeline from an S3 uploaded file.",
        "operationId": "process_document_api_pipeline_process_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineProcessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/jobs/{job_id}": {
      "get": {
        "tags": [
          "pipeline"
        ],
        "summary": "Get Job Status",
        "description": "Get pipeline job status and progress.",
        "operationId": "get_job_status_api_pipeline_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineJobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipeline/jobs/{job_id}/create-session": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Create Session From Pipeline",
        "description": "Create a learning session from a completed pipeline job.",
        "operationId": "create_session_from_pipeline_api_pipeline_jobs__job_id__create_session_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "max_steps",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 9999,
              "title": "Max Steps"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Service-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineSessionCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeEdgeResponse": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "title": "KnowledgeEdgeResponse"
      },
      "KnowledgeGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeNodeResponse"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges"
        ],
        "title": "KnowledgeGraphResponse"
      },
      "KnowledgeNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mastery": {
            "type": "number",
            "title": "Mastery"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "visited": {
            "type": "boolean",
            "title": "Visited"
          }
        },
        "type": "object",
        "required": [
          "id",
          "index",
          "name",
          "mastery",
          "status",
          "visited"
        ],
        "title": "KnowledgeNodeResponse"
      },
...
      "PipelineFailedBatchResponse": {
        "properties": {
          "batch_index": {
            "type": "integer",
            "title": "Batch Index"
          },
          "page_start": {
            "type": "integer",
            "title": "Page Start"
          },
          "page_end": {
            "type": "integer",
            "title": "Page End"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "batch_index",
          "page_start",
          "page_end",
          "reason"
        ],
        "title": "PipelineFailedBatchResponse"
      },
...
      "PipelineJobResultResponse": {
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/KnowledgeGraphResponse"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          },
          "n_concepts": {
            "type": "integer",
            "title": "N Concepts",
            "default": 0
          }
        },
        "type": "object",
        "title": "PipelineJobResultResponse"
      },
      "PipelineJobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "default": ""
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "queued",
              "loading",
              "chunking",
              "extracting",
              "embedding",
              "merging",
              "ranking",
              "verifying",
              "building_graph",
              "optimizing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step",
            "default": ""
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks",
            "default": 0
          },
          "page_batch_size": {
            "type": "integer",
            "title": "Page Batch Size",
            "default": 10
          },
          "batch_count": {
            "type": "integer",
            "title": "Batch Count",
            "default": 0
          },
          "failed_batch_count": {
            "type": "integer",
            "title": "Failed Batch Count",
            "default": 0
          },
          "partial_success": {
            "type": "boolean",
            "title": "Partial Success",
            "default": false
          },
          "concepts_extracted": {
            "type": "integer",
            "title": "Concepts Extracted",
            "default": 0
          },
          "concepts_after_merge": {
            "type": "integer",
            "title": "Concepts After Merge",
            "default": 0
          },
          "relations_verified": {
            "type": "integer",
            "title": "Relations Verified",
            "default": 0
          },
          "graph_stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Graph Stats"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "user_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "failed_batches": {
            "items": {
              "$ref": "#/components/schemas/PipelineFailedBatchResponse"
            },
            "type": "array",
            "title": "Failed Batches"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "is_terminal": {
            "type": "boolean",
            "title": "Is Terminal",
            "default": false
          },
          "is_delayed": {
            "type": "boolean",
            "title": "Is Delayed",
            "default": false
          },
          "created_at": {
            "type": "number",
            "title": "Created At",
            "default": 0.0
          },
          "updated_at": {
            "type": "number",
            "title": "Updated At",
            "default": 0.0
          },
          "heartbeat_at": {
            "type": "number",
            "title": "Heartbeat At",
            "default": 0.0
          },
          "retry_after_seconds": {
            "type": "integer",
            "title": "Retry After Seconds",
            "default": 2
          },
          "partial_graph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelinePartialGraphResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineJobResultResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "PipelineJobStatusResponse"
      },
      "PipelinePartialGraphEdgeResponse": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "title": "PipelinePartialGraphEdgeResponse"
      },
      "PipelinePartialGraphNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "PipelinePartialGraphNodeResponse"
      },
      "PipelinePartialGraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/PipelinePartialGraphNodeResponse"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/PipelinePartialGraphEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "title": "PipelinePartialGraphResponse"
      },
      "PipelineProcessResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "queued",
              "loading",
              "chunking",
              "extracting",
              "embedding",
              "merging",
              "ranking",
              "verifying",
              "building_graph",
              "optimizing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "page_batch_size": {
            "type": "integer",
            "title": "Page Batch Size",
            "default": 10
          },
          "retry_after_seconds": {
            "type": "integer",
            "title": "Retry After Seconds",
            "default": 2
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "filename",
          "file_size",
          "subject_id",
          "status",
          "status_url",
          "message"
        ],
        "title": "PipelineProcessResponse"
      },
      "PipelineSessionCreateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "n_concepts": {
            "type": "integer",
            "title": "N Concepts"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "active",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "n_concepts",
          "source",
          "job_id",
          "status"
        ],
        "title": "PipelineSessionCreateResponse"
      },
      "ProcessDocumentRequest": {
        "properties": {
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          },
          "prs_threshold": {
            "type": "number",
            "title": "Prs Threshold",
            "default": 0.75
          },
          "min_confidence": {
            "type": "number",
            "title": "Min Confidence",
            "default": 0.6
          },
          "apply_reduction": {
            "type": "boolean",
            "title": "Apply Reduction",
            "default": true
          },
          "page_batch_size": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Page Batch Size",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "file_url",
          "filename"
        ],
        "title": "ProcessDocumentRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "XUserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Authenticated user id forwarded by the frontend proxy."
      },
      "XServiceTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-service-token",
        "description": "Shared internal token used by the frontend proxy when calling the backend API."
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment behind the current API host."
    }
  ],
  "security": [
    {
      "XUserIdHeader": [],
      "XServiceTokenHeader": []
    }
  ]
}